+++ app/models/collection.rb
@@ -1,5 +1,7 @@
 class Collection < JupiterCore::LockedLdpObject
 
+  has_solr_exporter Exporters::Solr::CollectionExporter
+
   include ObjectProperties
 
   ldp_object_includes Hydra::Works::CollectionBehavior
@@ -7,23 +9,10 @@ class Collection < JupiterCore::LockedLdpObject
   # TODO: this should probably be renamed to share a name with member_of_paths on Item, so that their
   # facet results can be coalesced when Collections are mixed into search results along with Items, as in the
   # main search results
-  has_attribute :community_id, ::TERMS[:ual].path,
-                type: :path,
-                solrize_for: :pathing
-
-  has_attribute :description, ::RDF::Vocab::DC.description, solrize_for: [:search]
-  has_attribute :restricted, ::TERMS[:ual].restricted_collection, type: :boolean, solrize_for: :exact_match
-  has_multival_attribute :creators, ::RDF::Vocab::DC.creator, solrize_for: :exact_match
-
-  # TODO: refactor this next line and move the title into Fedora, if we're still on Fedora at that point.
-  #
-  # We got lucky in that there are not expected to be a large number of Collections in this phase of Jupiter
-  # but using +additional_search_index+ to store data that isn't recreatable solely by inspecting this object's
-  # Fedora record creates data-ordering issues that are complicated to work around during Solr-index recovery
-  # scenarios. See recover.rake for information on the particular problems this is causing and why we want to
-  # eliminate it.
-  additional_search_index :community_title, solrize_for: :sort,
-                                            as: -> { Community.find(community_id).title if community_id.present? }
+  has_attribute :community_id, ::TERMS[:ual].path
+  has_attribute :description, ::RDF::Vocab::DC.description
+  has_attribute :restricted, ::TERMS[:ual].restricted_collection
+  has_multival_attribute :creators, ::RDF::Vocab::DC.creator
 
   def community
     Community.find(community_id)