+++ app/models/exporters/solr/community_exporter.rb
@@ -1,0 +1,12 @@
+class Exporters::Solr::CommunityExporter < Exporters::Solr::BaseExporter
+
+  index :title, role: [:search, :sort]
+
+  # UAL attributes
+  index :fedora3_uuid, role: :exact_match
+  index :depositor, role: [:search]
+
+  index :description, role: [:search]
+  index :creators, role: :exact_match
+
+end